pyproject.toml
[project]
name = "currency-converter-using-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "streamlit>=1.43.2",
]
